* Use generics where possible * Use collection interfaces where possible * Misc cleanup

+++ client-java/trunk/src/main/org/firebirdsql/jdbc/field/FBIntegerField.java
@@ -137,7 +137,7 @@ class FBIntegerField extends FBField {
     }
     
     public void setShort(short value) throws SQLException {
-        setInteger((int)value);
+        setInteger(value);
     }
     
     public void setBoolean(boolean value) throws SQLException {
